backend/requirements.txt
fastapi
pymongo
passlib
pyjwt[crypto]
validate_email
bcrypt
git+https://gitea.ksol.io/karolyi/py3-validate-email@v1.0.9
